+++ fe/src/Component/OptionTabs/OptionsTabs.styled.tsx
